Fix search issue in scipy devdocs.
In https://scipy.github.io/devdocs, when you use the search box,
the links it returns include an "undefined" that make them invalid.
Example: first link of https://scipy.github.io/devdocs/search.html?q=eigh
is https://scipy.github.io/devdocs/generated/scipy.linalg.eighundefined
?highlight=eigh.

This is because the LINK_SUFFIX is not assigned any value. The fix is
to include the generated documentation_options.js file, so that
irrespective of the Sphinx version, the appropriate attributes
would be set.